something-something-prediction/requirements.txt
torch
torchvision
accelerate
diffusers[torch]
transformers
datasets
numpy
opencv-python
pillow
scikit-image
huggingface_hub
xformers
wandb
ftfy
tqdm